pyproject.toml
[project]
name = "talentlens"
version = "1.0.0"
description = "Resume screening application API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.27.1",
    "sqlalchemy==2.0.27",
    "alembic==1.13.1",
    "psycopg2-binary==2.9.9",
    "python-jose[cryptography]==3.3.0",
    "firebase-admin==6.4.0",
    "python-multipart==0.0.9",
    "aiofiles==23.2.1",
    "pydantic==2.6.1",
    "pydantic-settings==2.2.1",
    "httpx==0.26.0",
    "python-dotenv==1.0.1",
    "pydantic[email]",
    "google-generativeai>=0.5.4",
    "pypdf2>=3.0.0",
    "reportlab>=4.0.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py38"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.black]
line-length = 120
target-version = ["py38"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-ra -q --strict-markers"
asyncio_mode = "auto"